Activity feed improvements

+++ signingDayPro/Services/SDActivityFeedService.m
@@ -16,14 +16,6 @@
 #import "HTMLParser.h"
 #import "SDUtils.h"
 
-typedef enum {
-    SDUserTypePlayer = 1,
-    SDUserTypeTeam = 2,
-    SDUserTypeCoach = 3,
-    SDUserTypeHighSchool = 4,
-    SDUserTypeMember = 5
-} SDUserType;
-
 @interface SDActivityFeedService ()
 
 + (void)createCommentFromDictionary:(NSDictionary *)commentDictionary;
@@ -37,6 +29,7 @@ typedef enum {
 
 + (void)getActivityStoriesForUser:(User *)user
                          withDate:(NSDate *)date
+                  shouldDeleteOld:(BOOL)deleteOld
                  withSuccessBlock:(void (^)(NSDictionary *results))successBlock
                      failureBlock:(void (^)(void))failureBlock
 {
@@ -70,7 +63,7 @@ typedef enum {
                                 success:^(AFHTTPRequestOperation *operation, id JSON) {
                                     
                                     //no date provided that means we are downloading first page, old stories should be deleted
-                                    if (!date) {
+                                    if (!date && deleteOld) {
                                         [self markAllStoriesForDeletion];
                                     }
                                     
@@ -145,7 +138,7 @@ typedef enum {
                                     [context MR_saveToPersistentStoreAndWait];
                                     
                                     //returns only after deleting
-                                    if (!date) {
+                                    if (!date && deleteOld) {
                                         [self deleteAllMarkedStories];
                                     }
                                     if (successBlock) {
@@ -217,7 +210,6 @@ typedef enum {
     
     if (userTypeId == SDUserTypePlayer) {
         //user type: Player
-        user.userType = @"Player";
         NSDictionary *attributeDictionary = [dictionary objectForKey:@"Attributes"];
         
         if (attributeDictionary) {
@@ -231,7 +223,6 @@ typedef enum {
     }
     else if (userTypeId == SDUserTypeTeam) {
         //user type: TEAM
-        user.userType = @"Team";
         NSDictionary *attributeDictionary = [dictionary objectForKey:@"Attributes"];
         
         if (attributeDictionary) {
@@ -245,7 +236,6 @@ typedef enum {
     }
     else if (userTypeId == SDUserTypeCoach) {
         //user type: Coach
-        user.userType = @"Coach";
         NSDictionary *attributeDictionary = [dictionary objectForKey:@"Attributes"];
         
         if (attributeDictionary) {
@@ -257,7 +247,6 @@ typedef enum {
     else if (userTypeId == SDUserTypeHighSchool) {
  
         //user type: HighSchool
-        user.userType = @"HighSchool";
         NSDictionary *attributeDictionary = [dictionary objectForKey:@"Attributes"];
         
         if (attributeDictionary) {
@@ -268,12 +257,6 @@ typedef enum {
                 user.stateCode = [attributeDictionary valueForKey:@"StateCode"];
             }
         }
-    }
-    else if (userTypeId == SDUserTypeMember){
-        //user type: Member
-        user.userType = @"Member";
-        
-        //member doesn't have attributes so nothing to do here
     }
 }
